REFACTOR-002: extract automation invariants into ./parse/automation.ts
Pure structural move, no behavior change. Moves checkAutomationLane
(module-private, called from checkSongIntegrity) into its own module.

Third of the REFACTOR-002 stack; builds on the clip-invariants
extraction.

Refs #143
